refactor:(sc-test): 名前の修正とファイルの分割
