+++ crates/sc-test/Cargo.toml
@@ -1,5 +1,5 @@
 [package]
-name = "moro-test"
+name = "sc-test"
 version.workspace = true
 edition.workspace = true
 authors.workspace = true
@@ -7,5 +7,14 @@ repository.workspace = true
 rust-version.workspace = true
 
 [dependencies]
+anyhow = "1.0.89"
 tokio = { workspace = true, features = ["full"] }
 tokio-util = "0.7.12"
+
+[[example]]
+name = "01"
+path = "examples/01_local_spawn.rs"
+
+[[example]]
+name = "02"
+path = "examples/02_join.rs"
